Trying to render model

+++ include/render/PushConstantObject.h
@@ -7,8 +7,8 @@ namespace magmatic::render
 {
 	struct PushConstantObject
 	{
-		uint8_t texture_index;
-		uint8_t sampler_index;
+		int32_t texture_index;
+		int32_t sampler_index;
 		glm::mat4 matrix;
 	};
 }
